buflen is of size_t type. register MULTIAPPEND extension add MULTIAPPEND
* libbalsa/mailbox_imap.c: buflen is of size_t type.
* libbalsa/imap/imap-handle.[hc]: register MULTIAPPEND extension
* libbalsa/imap/imap-commands.[hc]: add MULTIAPPEND aware interface.
* libbalsa/imap/imap_tst.c: test it.

svn path=/trunk/; revision=7816

+++ libbalsa/mailbox_imap.c
@@ -2234,7 +2234,7 @@ get_section_for(LibBalsaMessage *msg, LibBalsaMessageBody *part)
 }
 struct part_data { char *block; unsigned pos; ImapBody *body; };
 static void
-append_str(unsigned seqno, const char *buf, int buflen, void *arg)
+append_str(unsigned seqno, const char *buf, size_t buflen, void *arg)
 {
     struct part_data *dt = (struct part_data*)arg;
